Main Files
the description will tell you about the use of the files in the program
email_sender:- the file that will actually be executed and need to be executed. uses:- taking credentials for the "first time setup" and then redirecting to run_run
run_run:- heart of the program. actually sends email
others :- detects if a file name "credentials.txt", if no then creates one and enters your username and password
convert:- used for extracting info from "credentials.txt" at runtime and converting it from a list to a string
Note:- the "credentials.txt" will be created on its own the first time the program is ran